src/main/java/payrollcasestudy/transactions/add/AddTimeCardTransaction.java
package payrollcasestudy.transactions.add;

import payrollcasestudy.entities.Employee;
import payrollcasestudy.PayrollDatabase;
import payrollcasestudy.entities.TimeCard;
import payrollcasestudy.entities.paymentclassifications.HourlyPaymentClassification;
import payrollcasestudy.entities.paymentclassifications.PaymentClassification;
import payrollcasestudy.transactions.Transaction;

public class AddTimeCardTransaction implements Transaction {

    private int date;
    private double hours;
    private int employeeId;

    public AddTimeCardTransaction(int date, double hours, int employeeId) {
        this.date = date;
        this.hours = hours;
        this.employeeId = employeeId;
    }

    @Override
    public void execute() {
        Employee employee = PayrollDatabase.globalPayrollDatabase.getEmployee(employeeId);
        if (employee != null){
            PaymentClassification paymentClassification = employee.getPaymentClassification();
            if (paymentClassification instanceof HourlyPaymentClassification){
                HourlyPaymentClassification classification = (HourlyPaymentClassification) paymentClassification;
                classification.addTimeCard(new TimeCard(date, hours));
            }
        }
    }
}
